+++ mpc4j-dp-service/src/main/java/edu/alibaba/mpc4j/dp/service/main/HhLdpMain.java
@@ -1,6 +1,7 @@
 package edu.alibaba.mpc4j.dp.service.main;
 
 import com.google.common.base.Preconditions;
+import edu.alibaba.mpc4j.common.tool.MathPreconditions;
 import edu.alibaba.mpc4j.common.tool.metrics.HeavyHitterMetrics;
 import edu.alibaba.mpc4j.common.tool.utils.PropertiesUtils;
 import edu.alibaba.mpc4j.dp.service.fo.FoLdpFactory;
@@ -102,6 +103,10 @@ public class HhLdpMain {
      */
     private final int testRound;
     /**
+     * if run plain
+     */
+    private final boolean plain;
+    /**
      * Frequency Oracle based type
      */
     private final List<FoLdpType> foTypeList;
@@ -121,7 +126,7 @@ public class HhLdpMain {
     public HhLdpMain(Properties properties) throws IOException {
         serverStopWatch = new StopWatch();
         clientStopWatch = new StopWatch();
-        reportFilePostfix = PropertiesUtils.readString(properties, "report_file_postfix");
+        reportFilePostfix = PropertiesUtils.readString(properties, "report_file_postfix", "");
         datasetName = PropertiesUtils.readString(properties, "dataset_name");
         // set dataset path
         datasetPath = PropertiesUtils.readString(properties, "dataset_path");
@@ -147,26 +152,19 @@ public class HhLdpMain {
                 .orElse(Integer.MAX_VALUE);
             dataStream.close();
         }
-        Preconditions.checkArgument(
-            domainMinValue < domainMaxValue,
-            "domain_min_value (%s) must be less than domain_max_value (%s)",
-            domainMinValue, domainMaxValue
-        );
+        MathPreconditions.checkLess("domain_min_value", domainMinValue, domainMaxValue);
         LOGGER.info("Domain Range: [{}, {}]", domainMinValue, domainMaxValue);
         domainSet = IntStream.rangeClosed(domainMinValue, domainMaxValue)
             .mapToObj(String::valueOf).collect(Collectors.toSet());
         int d = domainSet.size();
         // set heavy hitter
         k = PropertiesUtils.readInt(properties, "k");
-        Preconditions.checkArgument(k <= d, "k must be less than or equal to %s: %s", d, k);
+        MathPreconditions.checkLessOrEqual("k", k, d);
         // set privacy parameters
         double warmupPercentage = PropertiesUtils.readDouble(properties, "warmup_percentage");
-        Preconditions.checkArgument(
-            warmupPercentage > 0 && warmupPercentage < 1,
-            "warmup_percentage must be in range (0, 1): %s", warmupPercentage
-        );
+        MathPreconditions.checkPositiveInRange("warmup_percentage", warmupPercentage, 1.0);
         windowEpsilons = PropertiesUtils.readDoubleArray(properties, "window_epsilon");
-        alphas = PropertiesUtils.readDoubleArray(properties, "alpha");
+        alphas = PropertiesUtils.readDoubleArrayWithDefault(properties, "alpha");
         // set test round
         testRound = PropertiesUtils.readInt(properties, "test_round");
         // num and warmup num
@@ -174,13 +172,15 @@ public class HhLdpMain {
         int num = (int) dataStream.count();
         dataStream.close();
         warmupNum = (int) Math.round(num * warmupPercentage);
+        // set if run plain
+        plain = PropertiesUtils.readBoolean(properties, "plain", false);
         // set Frequency Oracle based types
-        String[] foTypeStrings = PropertiesUtils.readTrimStringArray(properties, "fo_types");
+        String[] foTypeStrings = PropertiesUtils.readTrimStringArrayWithDefault(properties, "fo_types");
         foTypeList = Arrays.stream(foTypeStrings)
             .map(FoLdpType::valueOf)
             .collect(Collectors.toList());
         // set HeavyGuardian based types
-        String[] hgTypeStrings = PropertiesUtils.readTrimStringArray(properties, "hg_types");
+        String[] hgTypeStrings = PropertiesUtils.readTrimStringArrayWithDefault(properties, "hg_types");
         hgTypeList = Arrays.stream(hgTypeStrings)
             .map(HhLdpType::valueOf)
             // ignore fo types
@@ -203,10 +203,32 @@ public class HhLdpMain {
         LOGGER.info("Correct heavy hitters: {}", correctHeavyHitters);
     }
 
+    String getReportFilePostfix() {
+        return reportFilePostfix;
+    }
+
+    boolean getPlain() {
+        return plain;
+    }
+
+    List<FoLdpType> getFoLdpList() {
+        return foTypeList;
+    }
+
+    List<HhLdpType> getHgTypeList() {
+        return hgTypeList;
+    }
+
+    double[] getAlphas() {
+        return alphas;
+    }
+
     public void run() throws IOException {
         // create report file
         LOGGER.info("Create report file");
-        String filePath = TASK_TYPE_NAME + "_" + datasetName + "_" + testRound + "_" + reportFilePostfix + ".txt";
+        String filePath = "".equals(reportFilePostfix)
+            ? TASK_TYPE_NAME + "_" + datasetName + "_" + testRound + ".txt"
+            : TASK_TYPE_NAME + "_" + datasetName + "_" + testRound + "_" + reportFilePostfix + ".txt";
         FileWriter fileWriter = new FileWriter(filePath);
         PrintWriter printWriter = new PrintWriter(fileWriter, true);
         // write tab
@@ -217,36 +239,43 @@ public class HhLdpMain {
             "            comm.(B)", "             mem.(B)",
             "                ndcg", "           precision", "                 abe", "                  re"
         );
-        runHeavyGuardian(printWriter);
-        if (!foTypeList.isEmpty()) {
-            for (double windowEpsilon : windowEpsilons) {
-                runFoHeavyHitter(windowEpsilon, printWriter);
+        if (plain) {
+            HhLdpAggMetrics heavyGuardianAggMetrics = runHeavyGuardian();
+            printInfo(printWriter, heavyGuardianAggMetrics);
+        }
+        for (double windowEpsilon : windowEpsilons) {
+            for (FoLdpType type : foTypeList) {
+                HhLdpAggMetrics foLdpAggMetrics = runFoHeavyHitter(type, windowEpsilon);
+                printInfo(printWriter, foLdpAggMetrics);
             }
         }
         if (hgTypeList.contains(HhLdpType.BASIC)) {
             for (double windowEpsilon : windowEpsilons) {
-                runBasicHgHeavyHitter(windowEpsilon, printWriter);
+                HhLdpAggMetrics basicHgLdpAggMetrics = runBasicHgHeavyHitter(windowEpsilon);
+                printInfo(printWriter, basicHgLdpAggMetrics);
             }
         }
         if (hgTypeList.contains(HhLdpType.ADV)) {
             for (double alpha : alphas) {
                 for (double windowEpsilon : windowEpsilons) {
-                    runAdvHhgHeavyHitter(windowEpsilon, alpha, printWriter);
+                    HhLdpAggMetrics advHgLdpAggMetrics = runAdvHhgHeavyHitter(windowEpsilon, alpha);
+                    printInfo(printWriter, advHgLdpAggMetrics);
                 }
             }
         }
         if (hgTypeList.contains(HhLdpType.RELAX)) {
             for (double windowEpsilon : windowEpsilons) {
-                runRelaxHhgHeavyHitter(windowEpsilon, printWriter);
+                HhLdpAggMetrics relaxHgLdpAggMetrics = runRelaxHhgHeavyHitter(windowEpsilon);
+                printInfo(printWriter, relaxHgLdpAggMetrics);
             }
         }
         printWriter.close();
         fileWriter.close();
     }
 
-    private void runHeavyGuardian(PrintWriter printWriter) throws IOException {
+    private HhLdpAggMetrics runHeavyGuardian() throws IOException {
         String typeName = " PURE_HG";
-        HhLdpAggMetrics aggMetrics = new HhLdpAggMetrics();
+        HhLdpAggMetrics aggMetrics = new HhLdpAggMetrics(typeName, null, null);
         for (int round = 0; round < testRound; round++) {
             HeavyGuardian heavyGuardian = new HeavyGuardian(1, k, 0);
             Stream<String> dataStream = StreamDataUtils.obtainItemStream(datasetPath);
@@ -278,10 +307,6 @@ public class HhLdpMain {
             // heavy hitter map
             Map<String, Double> heavyHitterMap = heavyGuardian.getRecordItemSet().stream()
                 .collect(Collectors.toMap(item -> item, item -> (double) heavyGuardian.query(item)));
-            Preconditions.checkArgument(
-                heavyHitterMap.size() == k,
-                "heavy hitter size must be equal to %s: %s", k, heavyHitterMap.size()
-            );
             // heavy hitter ordered list
             List<Map.Entry<String, Double>> heavyHitterOrderedList = new ArrayList<>(heavyHitterMap.entrySet());
             heavyHitterOrderedList.sort(Comparator.comparingDouble(Map.Entry::getValue));
@@ -300,30 +325,27 @@ public class HhLdpMain {
             metrics.setRe(HeavyHitterMetrics.relativeError(heavyHitterMap, correctCountMap));
             aggMetrics.addMetrics(metrics);
         }
-        // output report
-        printInfo(printWriter, typeName, null, null, aggMetrics);
+        return aggMetrics;
     }
 
-    private void runFoHeavyHitter(double windowEpsilon, PrintWriter printWriter) throws IOException {
-        for (FoLdpType foLdpType : foTypeList) {
-            HhLdpAggMetrics aggMetrics = new HhLdpAggMetrics();
-            for (int round = 0; round < testRound; round++) {
-                FoLdpConfig foLdpConfig = FoLdpFactory.createDefaultConfig(foLdpType, domainSet, windowEpsilon);
-                HhLdpConfig hhLdpConfig = new FoHhLdpConfig
-                    .Builder(foLdpConfig, k)
-                    .build();
-                HhLdpServer server = HhLdpFactory.createServer(hhLdpConfig);
-                HhLdpClient client = HhLdpFactory.createClient(hhLdpConfig);
-                HhLdpMetrics metrics = runLdpHeavyHitter(server, client);
-                aggMetrics.addMetrics(metrics);
-            }
-            printInfo(printWriter, "FO (" + foLdpType.name() + ")", windowEpsilon, null, aggMetrics);
+    private HhLdpAggMetrics runFoHeavyHitter(FoLdpType foLdpType, double windowEpsilon) throws IOException {
+        HhLdpAggMetrics aggMetrics = new HhLdpAggMetrics("FO (" + foLdpType.name() + ")", windowEpsilon, null);
+        for (int round = 0; round < testRound; round++) {
+            FoLdpConfig foLdpConfig = FoLdpFactory.createDefaultConfig(foLdpType, domainSet, windowEpsilon);
+            HhLdpConfig hhLdpConfig = new FoHhLdpConfig
+                .Builder(foLdpConfig, k)
+                .build();
+            HhLdpServer server = HhLdpFactory.createServer(hhLdpConfig);
+            HhLdpClient client = HhLdpFactory.createClient(hhLdpConfig);
+            HhLdpMetrics metrics = runLdpHeavyHitter(server, client);
+            aggMetrics.addMetrics(metrics);
         }
+        return aggMetrics;
     }
 
-    private void runBasicHgHeavyHitter(double windowEpsilon, PrintWriter printWriter) throws IOException {
+    HhLdpAggMetrics runBasicHgHeavyHitter(double windowEpsilon) throws IOException {
         HhLdpType type = HhLdpType.BASIC;
-        HhLdpAggMetrics aggMetrics = new HhLdpAggMetrics();
+        HhLdpAggMetrics aggMetrics = new HhLdpAggMetrics(HhLdpType.BASIC.name(), windowEpsilon, null);
         for (int round = 0; round < testRound; round++) {
             HgHhLdpConfig config = new HgHhLdpConfig
                 .Builder(type, domainSet, k, windowEpsilon)
@@ -333,12 +355,12 @@ public class HhLdpMain {
             HhLdpMetrics metrics = runLdpHeavyHitter(server, client);
             aggMetrics.addMetrics(metrics);
         }
-        printInfo(printWriter, type.name(), windowEpsilon, null, aggMetrics);
+        return aggMetrics;
     }
 
-    private void runAdvHhgHeavyHitter(double windowEpsilon, double alpha, PrintWriter printWriter) throws IOException {
+    HhLdpAggMetrics runAdvHhgHeavyHitter(double windowEpsilon, double alpha) throws IOException {
         HhLdpType type = HhLdpType.ADV;
-        HhLdpAggMetrics aggMetrics = new HhLdpAggMetrics();
+        HhLdpAggMetrics aggMetrics = new HhLdpAggMetrics(type.name(), windowEpsilon, alpha);
         for (int round = 0; round < testRound; round++) {
             HgHhLdpConfig config = new HgHhLdpConfig
                 .Builder(type, domainSet, k, windowEpsilon)
@@ -349,12 +371,12 @@ public class HhLdpMain {
             HhLdpMetrics metrics = runLdpHeavyHitter(server, client);
             aggMetrics.addMetrics(metrics);
         }
-        printInfo(printWriter, type.name(), windowEpsilon, alpha, aggMetrics);
+        return aggMetrics;
     }
 
-    private void runRelaxHhgHeavyHitter(double windowEpsilon, PrintWriter printWriter) throws IOException {
+    HhLdpAggMetrics runRelaxHhgHeavyHitter(double windowEpsilon) throws IOException {
         HhLdpType type = HhLdpType.RELAX;
-        HhLdpAggMetrics aggMetrics = new HhLdpAggMetrics();
+        HhLdpAggMetrics aggMetrics = new HhLdpAggMetrics(type.name(), windowEpsilon, null);
         for (int round = 0; round < testRound; round++) {
             HgHhLdpConfig config = new HgHhLdpConfig
                 .Builder(type, domainSet, k, windowEpsilon)
@@ -364,10 +386,10 @@ public class HhLdpMain {
             HhLdpMetrics metrics = runLdpHeavyHitter(server, client);
             aggMetrics.addMetrics(metrics);
         }
-        printInfo(printWriter, type.name(), windowEpsilon, null, aggMetrics);
+        return aggMetrics;
     }
 
-    private HhLdpMetrics runLdpHeavyHitter(HhLdpServer server, HhLdpClient client) throws IOException {
+    HhLdpMetrics runLdpHeavyHitter(HhLdpServer server, HhLdpClient client) throws IOException {
         // warmup
         AtomicInteger warmupIndex = new AtomicInteger();
         Stream<String> dataStream = StreamDataUtils.obtainItemStream(datasetPath);
@@ -426,10 +448,10 @@ public class HhLdpMain {
         return metrics;
     }
 
-    private void printInfo(PrintWriter printWriter, String type, Double windowEpsilon, Double alpha,
-                           HhLdpAggMetrics aggMetrics) {
-        String windowEpsilonString = windowEpsilon == null ? "-" : String.valueOf(windowEpsilon);
-        String alphaString = alpha == null ? "-" : String.valueOf(alpha);
+    private void printInfo(PrintWriter printWriter, HhLdpAggMetrics aggMetrics) {
+        String typeString = aggMetrics.getTypeString();
+        String windowEpsilonString = aggMetrics.getWindowEpsilonString();
+        String alphaString = aggMetrics.getAlphaString();
         double serverTime = aggMetrics.getServerTimeSecond();
         double clientTime = aggMetrics.getClientTimeSecond();
         long payloadBytes = aggMetrics.getPayloadBytes();
@@ -439,7 +461,7 @@ public class HhLdpMain {
         double abe = aggMetrics.getAbe();
         double re = aggMetrics.getRe();
         LOGGER.info("{}\t{}\t{}\t{}\t{}\t{}\t{}\t{}\t{}\t{}\t{}",
-            StringUtils.leftPad(type, 20),
+            StringUtils.leftPad(typeString, 20),
             StringUtils.leftPad(windowEpsilonString, 10),
             StringUtils.leftPad(alphaString, 10),
             StringUtils.leftPad(TIME_DECIMAL_FORMAT.format(serverTime), 20),
@@ -451,7 +473,7 @@ public class HhLdpMain {
             StringUtils.leftPad(DOUBLE_DECIMAL_FORMAT.format(abe), 20),
             StringUtils.leftPad(DOUBLE_DECIMAL_FORMAT.format(re), 20)
         );
-        printWriter.println(type + "\t" + windowEpsilonString + "\t" + alphaString + "\t"
+        printWriter.println(typeString + "\t" + windowEpsilonString + "\t" + alphaString + "\t"
             + serverTime + "\t" + clientTime + "\t" + payloadBytes + "\t" + memoryBytes + "\t"
             + ndcg + "\t" + precision + "\t" + abe + "\t" + re);
     }